requirements.txt
contourpy
cycler
fonttools
kiwisolver
matplotlib
numpy
packaging
pillow
pyneurosdk2
pyparsing
python-dateutil
six
pyserial